Bug 1828584 - Add JS_FrontendOnlyInit and JS_FrontendOnlyShutDown. r=bthrall
Differential Revision: https://phabricator.services.mozilla.com/D176512

+++ js/src/vm/Initialization.cpp
@@ -108,7 +108,7 @@ static void SetupCanonicalNaN() {
 extern "C" void install_rust_hooks();
 
 JS_PUBLIC_API const char* JS::detail::InitWithFailureDiagnostic(
-    bool isDebugBuild) {
+    bool isDebugBuild, FrontendOnly frontendOnly /* = FrontendOnly::No */) {
   // Verify that our DEBUG setting matches the caller's.
 #ifdef DEBUG
   MOZ_RELEASE_ASSERT(isDebugBuild);
@@ -131,11 +131,13 @@ JS_PUBLIC_API const char* JS::detail::InitWithFailureDiagnostic(
 
   PRMJ_NowInit();
 
-  // The first invocation of `ProcessCreation` creates a temporary thread
-  // and crashes if that fails, i.e. because we're out of memory. To prevent
-  // that from happening at some later time, get it out of the way during
-  // startup.
-  mozilla::TimeStamp::ProcessCreation();
+  if (frontendOnly == FrontendOnly::No) {
+    // The first invocation of `ProcessCreation` creates a temporary thread
+    // and crashes if that fails, i.e. because we're out of memory. To prevent
+    // that from happening at some later time, get it out of the way during
+    // startup.
+    mozilla::TimeStamp::ProcessCreation();
+  }
 
 #ifdef DEBUG
   CheckMessageParameterCounts();
@@ -143,7 +145,9 @@ JS_PUBLIC_API const char* JS::detail::InitWithFailureDiagnostic(
 
   SetupCanonicalNaN();
 
-  RETURN_IF_FAIL(js::TlsContext.init());
+  if (frontendOnly == FrontendOnly::No) {
+    RETURN_IF_FAIL(js::TlsContext.init());
+  }
 
 #if defined(DEBUG) || defined(JS_OOM_BREAKPOINT)
   RETURN_IF_FAIL(js::oom::InitThreadType());
@@ -169,13 +173,17 @@ JS_PUBLIC_API const char* JS::detail::InitWithFailureDiagnostic(
 
   js::coverage::InitLCov();
 
-  RETURN_IF_FAIL(js::jit::InitializeJit());
+  if (frontendOnly == FrontendOnly::No) {
+    RETURN_IF_FAIL(js::jit::InitializeJit());
+  }
 
   RETURN_IF_FAIL(js::InitDateTimeState());
 
+  if (frontendOnly == FrontendOnly::No) {
 #ifdef MOZ_VTUNE
-  RETURN_IF_FAIL(js::vtune::Initialize());
+    RETURN_IF_FAIL(js::vtune::Initialize());
 #endif
+  }
 
 #if JS_HAS_INTL_API
   if (mozilla::intl::ICU4CLibrary::Initialize().isErr()) {
@@ -183,22 +191,26 @@ JS_PUBLIC_API const char* JS::detail::InitWithFailureDiagnostic(
   }
 #endif  // JS_HAS_INTL_API
 
-  RETURN_IF_FAIL(js::CreateHelperThreadsState());
-  RETURN_IF_FAIL(FutexThread::initialize());
-  RETURN_IF_FAIL(js::gcstats::Statistics::initialize());
-  RETURN_IF_FAIL(js::InitTestingFunctions());
+  if (frontendOnly == FrontendOnly::No) {
+    RETURN_IF_FAIL(js::CreateHelperThreadsState());
+    RETURN_IF_FAIL(FutexThread::initialize());
+    RETURN_IF_FAIL(js::gcstats::Statistics::initialize());
+    RETURN_IF_FAIL(js::InitTestingFunctions());
+  }
 
   RETURN_IF_FAIL(js::SharedImmutableStringsCache::initSingleton());
   RETURN_IF_FAIL(js::frontend::WellKnownParserAtoms::initSingleton());
 
+  if (frontendOnly == FrontendOnly::No) {
 #ifdef JS_SIMULATOR
-  RETURN_IF_FAIL(js::jit::SimulatorProcess::initialize());
+    RETURN_IF_FAIL(js::jit::SimulatorProcess::initialize());
 #endif
 
 #ifndef JS_CODEGEN_NONE
-  // This is forced by InitializeJit.
-  MOZ_ASSERT(js::jit::CPUFlagsHaveBeenComputed());
+    // This is forced by InitializeJit.
+    MOZ_ASSERT(js::jit::CPUFlagsHaveBeenComputed());
 #endif
+  }
 
   libraryInitState = InitState::Running;
   return nullptr;
@@ -236,7 +248,9 @@ JS_PUBLIC_API bool JS::InitSelfHostedCode(JSContext* cx, SelfHostedCache cache,
   return true;
 }
 
-JS_PUBLIC_API void JS_ShutDown(void) {
+static void ShutdownImpl(JS::detail::FrontendOnly frontendOnly) {
+  using FrontendOnly = JS::detail::FrontendOnly;
+
   MOZ_ASSERT(
       libraryInitState == InitState::Running,
       "JS_ShutDown must only be called after JS_Init and can't race with it");
@@ -253,13 +267,15 @@ JS_PUBLIC_API void JS_ShutDown(void) {
   js::frontend::WellKnownParserAtoms::freeSingleton();
   js::SharedImmutableStringsCache::freeSingleton();
 
-  FutexThread::destroy();
+  if (frontendOnly == FrontendOnly::No) {
+    FutexThread::destroy();
 
-  js::DestroyHelperThreadsState();
+    js::DestroyHelperThreadsState();
 
 #ifdef JS_SIMULATOR
-  js::jit::SimulatorProcess::destroy();
+    js::jit::SimulatorProcess::destroy();
 #endif
+  }
 
   js::wasm::ShutDown();
 
@@ -278,19 +294,31 @@ JS_PUBLIC_API void JS_ShutDown(void) {
   mozilla::intl::ICU4CLibrary::Cleanup();
 #endif  // JS_HAS_INTL_API
 
+  if (frontendOnly == FrontendOnly::No) {
 #ifdef MOZ_VTUNE
-  js::vtune::Shutdown();
+    js::vtune::Shutdown();
 #endif  // MOZ_VTUNE
+  }
 
   js::FinishDateTimeState();
 
-  js::jit::ShutdownJit();
+  if (frontendOnly == FrontendOnly::No) {
+    js::jit::ShutdownJit();
+  }
 
   MOZ_ASSERT_IF(!JSRuntime::hasLiveRuntimes(), !js::WasmReservedBytes());
 
   js::ShutDownMallocAllocator();
 
   libraryInitState = InitState::ShutDown;
+}
+
+JS_PUBLIC_API void JS_ShutDown(void) {
+  ShutdownImpl(JS::detail::FrontendOnly::No);
+}
+
+JS_PUBLIC_API void JS_FrontendOnlyShutDown(void) {
+  ShutdownImpl(JS::detail::FrontendOnly::Yes);
 }
 
 JS_PUBLIC_API bool JS_SetICUMemoryFunctions(JS_ICUAllocFn allocFn,
